warp 2

+++ csrc/fused_softmax/scaled_upper_triang_masked_softmax.h
@@ -83,7 +83,7 @@ struct Max {
 };
 
 template <typename T>
-__device__ __forceinline__ T WARP_SHFL_XOR_NATIVE(T value, int laneMask, int width = warpSize, unsigned int mask = 0xffffffff)
+__device__ __forceinline__ T WARP_SHFL_XOR_NATIVE(T value, int laneMask, int width = warpSize, unsigned long long int mask = 0xffffffff)
 {
 #if defined(HIP_ENABLE_WARP_SYNC_BUILTINS)
     return __shfl_xor_sync(mask, value, laneMask, width);
